requirements.txt
pytest
selenium
webdriver-manager
allure-pytest
pytest-xdist
requests
faker